Changed functionality of offset parameter
Alternate fix for "removed items not on the path when offset is outside the -1, 1 boundary

Rather than removing instances, it loops them back to the start of the path, this method negates issuse from negative offset values by taking the absolute value of the offset. Reference issue #27 for issues with min/max values and modifier spin boxes.