Fix setup of use_chunks and chunk_dimensions ingame
Previously, these would only be updated if the script detected
that it was being ran in-editor, which allowed adjusting chunk
usage and dimensions fine on it. However, once ingame, the updated
values wouldn't apply, resulting in the defaults being used.

This commit fixes that, while also adding changes to make scatter
update whenever the chunk dimensions and the scatter item's visibility
ranges are altered, allowing for direct ingame and editor changes
without needing to alter another, potentially unrelated setting to
make scatter update.