Some improvements to ScatterCache
Previously the "Clear Cache" button had no function due to the cleared
cache not being saved to disk. This commit solves that.

Furthermore, saving was done on the main thread, which can cause
the editor to become unresponsive if there are many objects being
generated by Scatter nodes (files over 100MB were sure to freeze
the editor for a second or two), so now it is saved on a separate
thread unless the "dbg_disable_thread" variable is enabled... and
the same now applies to loading the scatter cache as well, which
did not load on the main thread if that was enabled previously.

Related to issue #238